requirements.txt
# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.32.0

# HTTP Client
requests==2.32.3
httpx==0.28.1

# File Upload
python-multipart==0.0.20

# AI Services
openai==1.109.1
google-generativeai==0.8.5
groq==0.32.0
huggingface_hub==1.1.4

# Database
pymongo==4.15.4

# Configuration
python-dotenv==1.1.1

# Utilities
pydantic==2.11.9
typing-extensions==4.12.2

# Video Processing (already in Dockerfile via ffmpeg)
# moviepy is not needed as we use ffmpeg directly